// src/app/interfaces/projectdetails.ts
export interface ProjectDetails {
    id: number;
    projectid: number;
    aspname: string;
    completeddate: Date;
    shift: string;
    currentstatus: string;
    e911completed: string;
    srscompleted: string;
    usedlongcable: string;
    dusasset: string;
    dusserial: string;
    dulasset: string;
    dulserial: string;
    xmuasset: string;
    xmuserial: string;
    installedserial: string;
    installedasset: string;
}